# Speed up `QM.add_variables_from()`