--- a/benchmarks/requirements.txt
+++ b/benchmarks/requirements.txt
@@ -1,3 +1,3 @@
 # we want the --strict flag which is not yet released as of October 2021
-git+https://github.com/airspeed-velocity/asv.git@7c938cf679d2e043829259f9f1061cf2564f7332
+asv==0.5.1
 virtualenv==16.7.12