```bash
Commit: f0413e35

qm.TimeAddVariablesFrom.time_add_variables_from_int [desktop/virtualenv-py3.8]         
  703±40μs                                                                                                                                       
  started: 2022-04-05 10:00:44, duration: 404ms                                                                                                  
                                                                                                                                                 
qm.TimeAddVariablesFrom.time_add_variables_from_str [desktop/virtualenv-py3.8]         
  2.05±0.2ms                                                                                                                                     
  started: 2022-04-05 10:00:44, duration: 404ms
```

```bash
Commit: c4bdc910 <0.11.x>

qm.TimeAddVariablesFrom.time_add_variables_from_int [desktop/virtualenv-py3.8]
  1.50±0.06ms
  started: 2022-04-05 10:03:15, duration: 404ms

qm.TimeAddVariablesFrom.time_add_variables_from_str [desktop/virtualenv-py3.8]
  2.72±0.1ms
  started: 2022-04-05 10:03:15, duration: 404ms
```