@arcondello 2022-04-05
Replaced by https://github.com/dwavesystems/dimod/pull/1162